+++ src/main/java/com/lann/openpark/charge/service/ChargeService.java
@@ -75,7 +75,7 @@ public class ChargeService {
         rule.setBorrowTimeConfig(borrowTimeConfig);
         LoopChargeConfig loopConfig = getLoopChargeConfigByJson(json.getJSONObject("loopConfig"));
         rule.setLoopConfig(loopConfig);
-        rule.setLimitCharge(json.getIntValue("limitCharge"));
+        rule.setLimitCharge(json.getDoubleValue("limitCharge"));
         AttachRule attachRule = getAttachRuleByJson(json.getJSONObject("attachRule"));
         rule.setAttachRule(attachRule);
 
@@ -146,6 +146,7 @@ public class ChargeService {
         }
         return config;
     }
+      
 
     private Rule getCaclRule(Rule rule, JSONObject json) {
         if (json == null) {
